requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
bleach==6.3.0
bokeh==3.6.3
cachetools==6.2.4
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
cloudpickle==3.1.2
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.2
cucim-cu13==25.12.0
cuda-bindings==13.1.1
cuda-core==0.3.2
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cuda-toolkit==13.1.0
cudf-cu13==25.12.0
cugraph-cu13==25.12.2
cuml-cu13==25.12.0
cupy-cuda13x==13.6.0
cuvs-cu13==25.12.0
cuxfilter-cu13==25.12.0
cycler==0.12.1
dask==2025.9.1
dask-cuda==25.12.0
dask-cudf-cu13==25.12.0
datashader==0.18.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distributed==2025.9.1
distributed-ucxx-cu13==0.47.0
executing==2.2.0
fastjsonschema==2.21.2
fastrlock==0.8.3
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
gast==0.6.0
geopandas==1.1.2
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
holoviews==1.20.2
idna==3.10
ImageIO==2.37.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.17.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
keras==3.9.2
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
libcudf-cu13==25.12.0
libcugraph-cu13==25.12.2
libcuml-cu13==25.12.0
libcuvs-cu13==25.12.0
libkvikio-cu13==25.12.0
libraft-cu13==25.12.0
librmm-cu13==25.12.0
libucx-cu13==1.19.0
libucxx-cu13==0.47.0
linkify-it-py==2.0.3
llvmlite==0.44.0
locket==1.0.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.1
msgpack==1.1.2
multidict==6.7.0
multipledispatch==1.0.0
namex==0.0.9
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.61.2
numba-cuda==0.19.1
numpy==2.1.3
nvidia-cublas==13.2.0.9
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cccl==13.1.78
nvidia-cuda-crt==13.1.80
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc==13.1.80
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc==13.1.80
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime==13.1.80
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft==12.1.0.31
nvidia-cufft-cu12==11.2.3.61
nvidia-curand==10.4.1.34
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver==12.0.7.41
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse==12.7.2.19
nvidia-cusparse-cu12==12.5.1.3
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.23.4
nvidia-nccl-cu13==2.28.9
nvidia-nvimgcodec-cu13==0.6.1.37
nvidia-nvjitlink==13.1.80
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvvm==13.1.80
nvtx==0.2.14
nx-cugraph-cu13==25.12.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
panel==1.7.5
param==2.3.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prometheus_client==0.23.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.23
pyct==0.6.0
Pygments==2.19.1
pylibcudf-cu13==25.12.0
pylibcugraph-cu13==25.12.2
pylibraft-cu13==25.12.0
pyogrio==0.12.1
pyparsing==3.2.3
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
pyzmq==26.4.0
raft-dask-cu13==25.12.0
rapids-dask-dependency==25.12.0
rapids-logger==0.2.3
referencing==0.37.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rmm-cu13==25.12.0
rpds-py==0.30.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.1.0
shapely==2.0.7
simpervisor==1.0.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.1
stack-data==0.6.3
statsmodels==0.14.4
tblib==3.2.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.12.20
tinycss2==1.4.0
toolz==1.1.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelite==4.6.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
ucxx-cu13==0.47.0
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xarray==2025.12.0
xyzservices==2025.11.0
yarl==1.22.0
zict==3.0.0
